# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/profile": {
      "get": {
        "description": "把 token 里的声明原样回显：`sub` + `username`。不含角色 / 权限 —— 本模块只回答\"你是谁\"，不回答\"你能不能\"。",
        "operationId": "AuthController_profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "token 有效",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProfileDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "入参校验失败（`errors[]` 里是字段级明细）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Bad Request",
                    "message": "Request validation failed",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f",
                    "errors": [
                      {
                        "field": "name",
                        "location": "body",
                        "message": "name must be longer than or equal to 2 characters"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证：缺少凭证、`Authorization` 不是 Bearer、凭证无效或已过期（响应头带 `WWW-Authenticate`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Unauthorized",
                    "message": "Missing or malformed credentials",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源或路由不存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Not Found",
                    "message": "user 999999 not found",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁（响应头带 `Retry-After` 与 `X-RateLimit-*`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Too Many Requests",
                    "message": "Too many requests, please try again later",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "500": {
            "description": "未捕获的内部错误（细节只进日志，不进响应）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Internal Server Error",
                    "message": "Internal server error",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          }
        },
        "summary": "当前身份（需要 Bearer token）",
        "tags": [
          "auth"
        ]
      }
    },
    "/validation-demo/users": {
      "post": {
        "operationId": "ValidationDemoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "入参校验失败（`errors[]` 里是字段级明细）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Bad Request",
                    "message": "Request validation failed",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f",
                    "errors": [
                      {
                        "field": "name",
                        "location": "body",
                        "message": "name must be longer than or equal to 2 characters"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源或路由不存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Not Found",
                    "message": "user 999999 not found",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "409": {
            "description": "业务冲突",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Conflict",
                    "message": "email neo@example.com already exists",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁（响应头带 `Retry-After` 与 `X-RateLimit-*`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Too Many Requests",
                    "message": "Too many requests, please try again later",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "500": {
            "description": "未捕获的内部错误（细节只进日志，不进响应）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Internal Server Error",
                    "message": "Internal server error",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          }
        },
        "summary": "创建用户（body 走全局校验管道）",
        "tags": [
          "validation-demo"
        ]
      },
      "get": {
        "description": "分页元信息在信封顶层的 `meta`（不是 `data.meta`）：`totalItems` / `itemsPerPage` / `currentPage`。",
        "operationId": "ValidationDemoController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页条数，超过 50 会被截断成 50（不是报错）",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序字段（只接受白名单里的列）",
            "schema": {
              "example": "id",
              "type": "string",
              "enum": [
                "id",
                "name",
                "email"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功（`data` 是这一页的数据数组）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "入参校验失败（`errors[]` 里是字段级明细）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Bad Request",
                    "message": "Request validation failed",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f",
                    "errors": [
                      {
                        "field": "name",
                        "location": "body",
                        "message": "name must be longer than or equal to 2 characters"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源或路由不存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Not Found",
                    "message": "user 999999 not found",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁（响应头带 `Retry-After` 与 `X-RateLimit-*`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Too Many Requests",
                    "message": "Too many requests, please try again later",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "500": {
            "description": "未捕获的内部错误（细节只进日志，不进响应）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Internal Server Error",
                    "message": "Internal server error",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          }
        },
        "summary": "分页查询用户",
        "tags": [
          "validation-demo"
        ]
      }
    },
    "/validation-demo/users/{id}": {
      "get": {
        "operationId": "ValidationDemoController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户 id，从 1 开始",
            "schema": {
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "入参校验失败（`errors[]` 里是字段级明细）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Bad Request",
                    "message": "Request validation failed",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f",
                    "errors": [
                      {
                        "field": "name",
                        "location": "body",
                        "message": "name must be longer than or equal to 2 characters"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源或路由不存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Not Found",
                    "message": "user 999999 not found",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁（响应头带 `Retry-After` 与 `X-RateLimit-*`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Too Many Requests",
                    "message": "Too many requests, please try again later",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "500": {
            "description": "未捕获的内部错误（细节只进日志，不进响应）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Internal Server Error",
                    "message": "Internal server error",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          }
        },
        "summary": "查询单个用户",
        "tags": [
          "validation-demo"
        ]
      },
      "patch": {
        "description": "`UpdateUserDto` 由 `PartialType(CreateUserDto)` 生成：所有字段可选，但保留原校验规则。",
        "operationId": "ValidationDemoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户 id",
            "schema": {
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功（只覆盖传入的字段）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "入参校验失败（`errors[]` 里是字段级明细）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Bad Request",
                    "message": "Request validation failed",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f",
                    "errors": [
                      {
                        "field": "name",
                        "location": "body",
                        "message": "name must be longer than or equal to 2 characters"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源或路由不存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Not Found",
                    "message": "user 999999 not found",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "409": {
            "description": "业务冲突",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Conflict",
                    "message": "email neo@example.com already exists",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁（响应头带 `Retry-After` 与 `X-RateLimit-*`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Too Many Requests",
                    "message": "Too many requests, please try again later",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "500": {
            "description": "未捕获的内部错误（细节只进日志，不进响应）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Internal Server Error",
                    "message": "Internal server error",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          }
        },
        "summary": "局部更新用户",
        "tags": [
          "validation-demo"
        ]
      }
    },
...
    "/validation-demo/boom": {
      "get": {
        "description": "抛普通 Error ⇒ HTTP 500 + 通用文案；完整堆栈只出现在服务端日志里，并且与响应头 `x-request-id` 是同一个 `traceId`。",
        "operationId": "ValidationDemoController_boom",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "入参校验失败（`errors[]` 里是字段级明细）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Bad Request",
                    "message": "Request validation failed",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f",
                    "errors": [
                      {
                        "field": "name",
                        "location": "body",
                        "message": "name must be longer than or equal to 2 characters"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源或路由不存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Not Found",
                    "message": "user 999999 not found",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁（响应头带 `Retry-After` 与 `X-RateLimit-*`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Too Many Requests",
                    "message": "Too many requests, please try again later",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "500": {
            "description": "未捕获的内部错误（细节只进日志，不进响应）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Internal Server Error",
                    "message": "Internal server error",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          }
        },
        "summary": "故意触发 500（堆栈只进日志、不进响应）",
        "tags": [
          "validation-demo"
        ]
      }
    },
...
  },
  "info": {
    "title": "nestjs-start API",
    "description": "`nestjs-start` 的接口文档。\n\n**响应契约**：\n- 成功 `{ success: true, data, meta? }`\n- 失败 `{ success: false, error, message, traceId?, errors? }`\n\n`success` 是唯一判据；`code` 是机器判据（**不要 parse `message`**）；\n数字状态码只在 HTTP 状态行里；`traceId` 同时回写在 `x-request-id` 响应头。\n详见 `docs/validation.md` §9。",
    "version": "1.0.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发"
    }
  ],
  "components": {
    "schemas": {
...
      "ProfileDto": {
        "type": "object",
        "properties": {
          "sub": {
            "type": "string",
            "description": "主体 id（JWT 的 `sub`）",
            "example": "1"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "neo"
          }
        },
        "required": [
          "sub",
          "username"
        ]
      },
      "ResponseEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "description": "handler 的返回值（分页时是这一页的数据数组）"
          },
          "meta": {
            "description": "仅分页成功时出现",
            "type": "object",
            "required": [
              "totalItems",
              "itemsPerPage",
              "currentPage"
            ],
            "properties": {
              "totalItems": {
                "type": "integer",
                "example": 3
              },
              "itemsPerPage": {
                "type": "integer",
                "example": 50
              },
              "currentPage": {
                "type": "integer",
                "example": 1
              }
            }
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Request validation failed"
          },
          "traceId": {
            "description": "请求 id：同时回写在 `x-request-id` 响应头、并写进服务端日志，用来把一次响应和日志对上。",
            "type": "string",
            "example": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
          },
          "errors": {
            "description": "仅校验类失败才有（`field` 是权威定位，嵌套用点号）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
...
    }
  }
}